+++ example/.flutter-plugins-dependencies
@@ -1,0 +1,1 @@
+{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"flutter_uploader","path":"/Users/sebastian/git/ened.flutter_uploader/","dependencies":[]},{"name":"image_picker","path":"/Users/sebastian/dev/flutter/.pub-cache/hosted/pub.dartlang.org/image_picker-0.6.7+4/","dependencies":[]},{"name":"path_provider","path":"/Users/sebastian/dev/flutter/.pub-cache/hosted/pub.dartlang.org/path_provider-1.6.11/","dependencies":[]}],"android":[{"name":"flutter_plugin_android_lifecycle","path":"/Users/sebastian/dev/flutter/.pub-cache/hosted/pub.dartlang.org/flutter_plugin_android_lifecycle-1.0.8/","dependencies":[]},{"name":"flutter_uploader","path":"/Users/sebastian/git/ened.flutter_uploader/","dependencies":[]},{"name":"image_picker","path":"/Users/sebastian/dev/flutter/.pub-cache/hosted/pub.dartlang.org/image_picker-0.6.7+4/","dependencies":["flutter_plugin_android_lifecycle"]},{"name":"path_provider","path":"/Users/sebastian/dev/flutter/.pub-cache/hosted/pub.dartlang.org/path_provider-1.6.11/","dependencies":[]}],"macos":[{"name":"path_provider_macos","path":"/Users/sebastian/dev/flutter/.pub-cache/hosted/pub.dartlang.org/path_provider_macos-0.0.4+3/","dependencies":[]}],"linux":[{"name":"path_provider_linux","path":"/Users/sebastian/dev/flutter/.pub-cache/hosted/pub.dartlang.org/path_provider_linux-0.0.1+2/","dependencies":[]}],"windows":[],"web":[]},"dependencyGraph":[{"name":"flutter_plugin_android_lifecycle","dependencies":[]},{"name":"flutter_uploader","dependencies":[]},{"name":"image_picker","dependencies":["flutter_plugin_android_lifecycle"]},{"name":"path_provider","dependencies":["path_provider_macos","path_provider_linux"]},{"name":"path_provider_linux","dependencies":[]},{"name":"path_provider_macos","dependencies":[]}],"date_created":"2020-07-15 19:29:03.568760","version":"1.17.5"}
