+++ example/android/app/src/main/AndroidManifest.xml
@@ -18,11 +18,11 @@
     android:networkSecurityConfig="@xml/network_security_config"
     tools:ignore="GoogleAppIndexingWarning">
     <activity
-      android:name=".MainActivity"
-      android:configChanges="orientation|keyboardHidden|keyboard|screenSize|locale|layoutDirection|fontScale|screenLayout|density|uiMode"
-      android:hardwareAccelerated="true"
+      android:name="io.flutter.embedding.android.FlutterActivity"
       android:launchMode="singleTop"
       android:theme="@style/LaunchTheme"
+      android:configChanges="orientation|keyboardHidden|keyboard|screenSize|locale|layoutDirection|fontScale|screenLayout|density|uiMode"
+      android:hardwareAccelerated="true"
       android:windowSoftInputMode="adjustResize">
       <!-- This keeps the window background of the activity showing
            until Flutter renders its first frame. It can be removed if
@@ -32,10 +32,17 @@
         android:name="io.flutter.app.android.SplashScreenUntilFirstFrame"
         android:value="true" />
       <intent-filter>
-        <action android:name="android.intent.action.MAIN" />
-        <category android:name="android.intent.category.LAUNCHER" />
+        <action android:name="android.intent.action.MAIN"/>
+        <category android:name="android.intent.category.LAUNCHER"/>
       </intent-filter>
     </activity>
+
+    <activity
+      android:name=".EmbeddingV1Activity"
+      android:theme="@style/LaunchTheme"
+      android:configChanges="orientation|keyboardHidden|keyboard|screenSize|locale|layoutDirection|fontScale"
+      android:hardwareAccelerated="true"
+      android:windowSoftInputMode="adjustResize" />
 
     <provider
       android:name="androidx.work.impl.WorkManagerInitializer"
@@ -54,5 +61,7 @@
         android:name="com.bluechilli.flutteruploader.UPLOAD_CONNECTION_TIMEOUT_IN_SECONDS"
         android:value="3600" />
     </provider>
+
+    <meta-data android:name="flutterEmbedding" android:value="2"/>
   </application>
 </manifest>
