+++ example/android/app/src/main/java/com/bluechilli/flutteruploaderexample/EmbeddingV1Activity.java
@@ -1,13 +1,14 @@
 package com.bluechilli.flutteruploaderexample;
 
 import android.os.Bundle;
+import com.bluechilli.flutteruploader.FlutterUploaderPlugin;
 import io.flutter.app.FlutterActivity;
-import io.flutter.plugins.GeneratedPluginRegistrant;
 
-public class MainActivity extends FlutterActivity {
+public class EmbeddingV1Activity extends FlutterActivity {
+
   @Override
   protected void onCreate(Bundle savedInstanceState) {
     super.onCreate(savedInstanceState);
-    GeneratedPluginRegistrant.registerWith(this);
+    FlutterUploaderPlugin.registerWith(registrarFor("com.bluechilli.flutteruploader"));
   }
 }
